pyits/pyproject.toml
[project]
name = "pyits"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Krzysztof Rusek", email = "krusek@agh.edu.pl" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = ["ref_gsd>=0.1.0","pandas"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[tool.hatch.version]
path = "pyits/__init__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest","seaborn","jaxlib>=0.4.18"
]
path='./venv'

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"


